+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Fixed
+- (cmap) Incorrectly returning glyph ID `0` instead of `None` for format 0
 
 ## [0.8.1] - 2020-07-29
 ### Added